Consider empty logs for tap tests as legal
A previous CL made it so that if a robot gesture returned no events the
test would continue on instead of halting.  This is great, but in the
case of "tap" tests, it's not exactly what we'd like.  For the tap
tests, we're trying to see *if* the pad saw the finger, so when the pad
fails to register a finger, we don't want to just skip it, we want to
consider it a failure and save that result.

This CL adds an exception to the validators that allows empty logs to
get validated if it's a robot test that's doing taps.

BUG=chromium:371505
TEST=I ran missed tap tests and the tracking ID count validator was
effected instead of simply skipping the log.  On other gestures that
failed (non-tap) the gesture was skipped outright and didn't effect the
results other than a blank placeholder that was put in the html report
which is nice actually as it acts as a reminder that the test wasn't run.

Change-Id: If0670a3550f4e1bc272d4139ba146a8e4ce416be
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/200095
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>