autotest (network): make network_DefaultProfileCreation more robust
Previously, this test would fail if the default profile was not
present before the test ran. While that shouldn't normally happen,
checking that case isn't the purpose of this test. (The purpose
of this test is to see what happens when we start shill anew.)

Update the test so that it can pass even if there is no default
profile at the time the test starts.

While there: fix inconsitent whitespace in existing code.

BUG=chromium:332767
TEST=manual (see below)

1. run test: passes
2. pass a directory to os.remove(). test fails.
3. pass a non-existent path to os.remove(). test passes.

Change-Id: I0739e4577bc29669c9a789744e9bef310d446d45
Reviewed-on: https://chromium-review.googlesource.com/201210
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Mukesh Agrawal <quiche@google.com>