Tool to collect a sequence of images from video.
Verifying that a video does not have glitches on a certain build we
need to have copies of reference/golden images that are known to be good.

This CL brings a stand-alone tool that can be run on a particular device
to capture images at specified time stamps in a video.

  ** Removed e.g:mp4 in commnad line argument  help text since I now provide mp4 as a choice.

BUG=chromium:363815
TEST=Ran tool on a LINK device and verified images were captured.

Change-Id: I1ab5c09ab2bf5aa6ecd0c408a7e031427f992f46
Reviewed-on: https://chromium-review.googlesource.com/201773
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>