Revert "Remove autotest install from WiFiClient"
WiFiClient depends on the XMLRPC server script being installed on the
DUT.  Unfortunately, in server tests, this only happens if we install it
explicitly or if the script is left over from some previous test.  We'll
have to fix the issue this patch was meant to correct in some other way.

TEST=It used to work this way.
BUG=chromium:378814

This reverts commit 21d238484e0594ac832ce78d54345924602774c7.

Change-Id: I266f8c19f694feb784527e16d63a71a7bc91d468
Reviewed-on: https://chromium-review.googlesource.com/202043
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>