Increase the number of expect tracking IDs
The robot performs 50 taps in the "stationary finger with taps around
it" test.  Previously the test only validated that there were >=2
tracking IDs, which would indeed be bad if there was only 1, but since
there should be 51, it's a little lax in the situation where the taps
are getting missed.  This CL sets it to 25 instead.  This test is quite
hard, so even good touchpads will miss a few tracking IDs (hence 25
instead of 50) and this gives them a little wiggle-room.  From the
performance I've seen on other devices, this seems like a reasonable
cutoff.

BUG=chromium:378846
TEST=none, this is just a simple numeric change

Change-Id: I31e95eabfe5fafad228166d8d525b17148840134
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202067