Add default values for start and stop time.
The tool to collect golden images for a given video required user
to specify the start time and stop time.

This CL uses 1st second as the default start time to capture screenshots
and the length of the video as the default stop time to capture screenshots.

BUG=chromium:363815
TEST=Ran tool as follows: python collect_golden_images.py mp4 480p 120
Ensured the capturing started at 1st second and went all the max value as per the interval.

Change-Id: Icd9bb1d6251b55a60a94e05950da5d8240569f67
Reviewed-on: https://chromium-review.googlesource.com/202075
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>