Update the robot wrapper to use new reset script
There is a patch up (that this CL depends on) that adds some additional
functionality to the touchbotII "reset.py" script and changes the
command line parameters for it.  This allows the touchbot wrapper to
call it to move safely to the nest as well as to the pad, and this CL
takes advantage of that.

BUG=chromium:378859
TEST=manually ran the patched robot test and it worked
CQ-DEPEND=I39ae10afb60e4d22153af230d8254ff21545159f

Change-Id: I0b566b6c1b59f59bdab3aa0d18b5f5b60bd58610
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202125
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>