Delay between setting hostevent and pressing power button
This adds a short delay (200ms) in enable_rec_mode_and_reboot
between setting the hostevent flag for recovery mode and pressing
the power button. This is intended to give the EC enough time
to receive and process the command, which may be especially long
if buffered DMA input is enabled.

BUG=chrome-os-partner:29210
TEST=Big boots into recovery mode more reliably.
Signed-off-by: David Hendricks <dhedndrix@chromium.org>

Change-Id: I006bf15e01e73f25a0295d2bbd20ca88493d2d24
Reviewed-on: https://chromium-review.googlesource.com/202225
Commit-Queue: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Vic Yang <victoryang@chromium.org>
Tested-by: Vic Yang <victoryang@chromium.org>