[autotest] move django.sessions cleanup from 24 hours to 5 mins.
Currently on Moblab our database's ibdata1 file is hitting its max
size causing all sorts of database problems. We suspect this is due
to the lack of cleanup of the django.sessions table and are now going
to run this as part of the 5 min cleanup vs the 24 hour cleanup.

BUG=chromium:379199
TEST=Schedule is running on fresh moblab image.

Change-Id: Ibd5261591b5edebd7cafb46a311e2513c1ca713e
Reviewed-on: https://chromium-review.googlesource.com/202317
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>