service_stopper: Enable context manager usage and add usage instructions
In the offchance that one wishes to start and stop services in each
iteration of the test, it makes more sense to use a context manager
style of using this class.  However, in most cases, one should likely
use initialize/cleanup, so this has been included as part of the
documentation for the file.

TEST=None
BUG=None

Change-Id: Ic8886438744b73360dbc91baf704e1d666cb3e04
Reviewed-on: https://chromium-review.googlesource.com/202354
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>