Chameleon: Remove the dependency on Xlib and use xdotool instead
Recently, the python-xlib was removed from the test image. Some of the
Chameleon tests depend on this library to toggle mirror mode by emulating
a Ctrl+F4.

This change is to remove the dependency on Xlib and uses the xdotool
(an X automation tool) for key emulation.

BUG=chromium:379853
TEST=Ran the Chameleon tests display_Resolution and display_HotPlugAtSuspend
on Peppy and verified they passed.

Change-Id: I56b3c061f3e6f72ea7ce39bb810efa0160a82d6c
Reviewed-on: https://chromium-review.googlesource.com/202364
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>