Remove possibility of running the config test 2x.
In the current retry flow it is possible that the same config
failure test will run test.  That means there will be two test
folders with the same name which causes the test to abort.

I also moved clearing out the command list to the finally
in the ap_cartridge.  If screenshot capturing throws an
exception, the command list won't be cleared.  Then when we
attempt to power down the AP it may perform other actions
that will throw an excpetion in the runner.  That exception
will abort the test.

BUG=None
TEST=Manual; Ran connect/disconnect

Change-Id: I96096ece08add659e4ee8ced0038bdeaf88711a3
Reviewed-on: https://chromium-review.googlesource.com/202447
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>