glbench: add --notemp option
The temperature checking of glbench is really good in the lab when running
glbench as part of a test suite.

However, it is very annoying when running glbench manually.

The --hasty option turns off temperature checking, but it also reduces the
number of iterations and tests, which makes it much less useful.

Add a --notemp option to just turn off the temperature checking.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=glbench --notemp

Change-Id: Ieeb26f0ab184948db6c80ce9dcea66e8a8735e7c
Reviewed-on: https://chromium-review.googlesource.com/202476
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>