Put video.html in the correct folder.
Video glitch detection tests failed in the lab because the page was
unable to load in the specified time (5s).
This is because video.html was not placed in the correct directory.

This CL add a change to copy video.html to whatever the binary directory
that autotest gave us at run time.

Local tests passed because there was a copy of video.html in the test case
directory that must have been the one used.

BUG=chromium:363815
TEST=Removed file video.html that was lurking in the test case directory.
Ran tests with change. Tests ran correctly.

Change-Id: Ia60a37b04e466e3a85d20c7f0822d5d3af8c1820
Reviewed-on: https://chromium-review.googlesource.com/202541
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>