Update RPC server shutdown timeout.
When shutting down RPC server, it will restart the chrome UI and wait for the
login prompt to comeback before exiting. This is to make sure chrome UI
is restored to a good state when the test exit.

We currently have timeout of 20 seconds for RPC server to shutdown, and there
are some cases where chrome devices took more than 20 seconds to restart the UI.
To work around this problem, the RPC server shutdown timeout is increased to
90 seconds, which will be the same as the restart UI timeout.

BUG=None
TEST=Run network_WiFi_SimpleConnect.wifi_check11a test

Change-Id: I00a4898fe1a299feb7e4f55f4a6b89389f0293d6
Reviewed-on: https://chromium-review.googlesource.com/202716
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>
Tested-by: Peter Qiu <zqiu@chromium.org>