cellular: Map assertion ids to descriptions in MBIMComplianceAssertionError.
This CL streamlines the way assertion errors are raised in the MBIM compliance
suite by having the class accept an assertion id, and mapping it internally to
the actual assertion error string.

Additionally, a subclass is provided to raise assertion errors that don't fit in
with the MBIM compliance doc list of assertions.

BUG=chromium:376958
TEST=python mbim_errors_unittest.py

Change-Id: I68c2d1262aa4367c0d6f9d981c3a4bb74e91aaee
Reviewed-on: https://chromium-review.googlesource.com/202718
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>