Collect system tray screenshots.
We are collecting system tray screenshots so that we can observe
how the system tray looks like in the lab.

We eventually intend to use image comparison based to tests to detect
basic rendering flaws in the system tray.

However, it is difficult to establish a baseline with tray as we don't
know a good state that the tray starts in.

This test won't fail, but it will give us the ability to look at the system
tray across many builds.

** Moved pixel measurements to be inputs from control file. Renamed control file accordingly.

** capitalized and moved out of class declaration variables that are meant
to be constants indicating that they are not meant to be changed.

** Added dependencies flag in the control file to point to the link board.
Consequently, removed code to throw on other boards.

** We will only run this test on link for now. But we need to investigate
the possibility of specifying exactly which boards we are interested in.

BUG=chromium:374981
TEST=ran a local test on link and opened biopic link to see the image.

Change-Id: Ifa2591604d019efe8539c061a32d4a0b24505f8a
Reviewed-on: https://chromium-review.googlesource.com/202825
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>