glbench: rename no_fill_compositing to compositing_no_fill
All of the other glbench test names follow the pattern <test>_<condition>.
This one test does <condition>_<test>.

It would be nice if we can switch it.
Luckily, this test does not draw anything, so there are no reference files
to rename.

However, this does make it harder to compare to historical data.  I'm not
so sure that matters for this test, however.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=test_that graphics_GLBench

Change-Id: Iebf41157d713ecb6968b8d22a7b9f1bb47c5bb13
Reviewed-on: https://chromium-review.googlesource.com/202883
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>