Add a script to find results not offloaded by gs_offloader.
When gs_offloader fails to offload results, a common failure
message looks like this:
    CommandException: 1 file/object could not be transferred.
The error output doesn't really explain what file failed to upload.

This adds a script to compare the list of files for a job results
directory with the files already uploaded.  The output of the
script is a list of files in the results directory that aren't
yet in Google Storage.

BUG=chromium:220050
TEST=Run it on atlantis2 on a recently failed job directory

Change-Id: I3a0c44c543d5616086dd336a9d78764979891b29
Reviewed-on: https://chromium-review.googlesource.com/203014
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>