Bluetooth SDP autotest: tests for attributes of fake service
This CL adds tests for those attributes, which are not present in any
of existing services. In order to do this, DUT starts a fake service,
which contains all this attributes with some fake value. After that
tester sends requests to the DUT and ensures that responses are
correct.

BUG=chromium:329044
TEST=test_that --board ${BOARD} ${HOSTNAME} bluetooth_SDP_ServiceAttributeRequest

Change-Id: Ica66c77e9fe71d6c179b88a648bd8aa5b0315871
Reviewed-on: https://chromium-review.googlesource.com/203190
Tested-by: Artem Rakhov <arakhov@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Artem Rakhov <arakhov@chromium.org>