Make sure items are removed from the locked list.
If unlock_all_aps is called everything works.  However if
unlock_one_ap is called from another method, items remain in the
locked list.  The result is a list of errors thrown when we try
to unlock them again.

The unit test has drifted too far from the current design:
  - new APSpec
  - new reclaim functionality

so I have removed it.  I am not sure how much value there is with
creating a new one since the best test is just to run against one
AP in the chaos lab.

Also verify the regex return for the firmware version is not
None.

BUG=chromium:384765
TEST=Manual; Ran it

Change-Id: Icf3611891ad24ade3a86851ec3afc3c6404b01f1
Reviewed-on: https://chromium-review.googlesource.com/203855
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Tien Chang <tienchang@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>