autotest (network): update network_UdevRename test
This test hasn't been used in a while, but I think it's still useful.
Bring it up to date, by a) adding mlan as a prefix for wireless devices,
and b) fixing how the test reads the name of the network device's
kernel module.

Also, add this test to the network_nightly suite.

While there: bring the control file in line with current style.

BUG=chromium:384147
TEST=ran this test against a peppy

Change-Id: I7b56011fee34ce2fd1e5893a9e2a21e863da39eb
Reviewed-on: https://chromium-review.googlesource.com/203746
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>