[autotest] Fix apache conf rewrite rules to only match the beginning of urls.
In apache conf files, most of the rewrite rules start with "/" without
an anchor (^). This will results in matching any part of the urls instead of
matching the beginning of urls.

For example, the rule "/tko(.*)" will match "/new_tko/tkoclient.css", which
is not intended to match it.

I added anchors to all the rewrite rules designed to match the beginning
of urls.

BUG=chromium:244451
DEPLOY=apache
TEST=ran afe and inspected the loading of static files through browser console

Change-Id: I13116e69b5273bf20c7323c720f051cdad71a853
Reviewed-on: https://chromium-review.googlesource.com/203659
Tested-by: Jiaxi Luo <jiaxiluo@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Jiaxi Luo <jiaxiluo@chromium.org>