autoupdate_EndToEndTest: Collect and report performance data.
While performing an update, run a script on the DUT to measure how
many resources update_engine and its helpers are using. For now we
only record the peak memory usage but the performance monitoring
script can easily be extended to record other things.

For now, we report all samples under the same name which means that
multiple samples from the same source version are reported (one for
the full update and one for each delta). In the future we might want
to report this separately.

BUG=chromium:385305
TEST=Ran the autoupdate_EndToEndTest without any problems.

Change-Id: I51f24e11d6dc28a55807440b116ddfc32b170152
Reviewed-on: https://chromium-review.googlesource.com/204160
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>