Clean up network entries in network_WiFi_SimpleConnect
This logic removes entries from the test profile after each
attempt to connect to a new network.  The pleasant side effect
of cleaning things as we go along is that we can't accidentally
attempt to scan for multiple hidden networks from the past
and miss the one we're interested in seeing.

TEST=network_WiFi_SimpleConnect.wifi_checkHidden passes.
BUG=chromium:382761

Change-Id: I04dc83f3a82c40a6760afc78c78795bab73ceb90
Reviewed-on: https://chromium-review.googlesource.com/204429
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>