glbench: allow tests to skip MD5 / image save step
Several glbench test cases don't draw anything to the framebuffer, so
reading back pixels, computing an MD5 hash on them, and saving them as a
.png is just wasted effort.

Furthermore, every time a test case is added that doesn't need this pixel
verification must also be added to the graphics_GLBench no_checksum_tests
list.

It's much easier to just let the test request that this stage be skipped,
and make graphics_GLBench just skip verification if no .png is provided.
When a test doesn't write a .png, the "imagefile" field is set to "[none]".
This is nominally done so that graphics_GLBench can distinguish this case
from a test that has 'value=0', which already uses the NULL imagefile
("[]").  Although, in truth, it is done because graphics_GLBench currently
treats NULL "[]" imagefile as a "knownbad" image, since the
glbench_knownbad_images.txt ends with a blank line.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=test_that graphics_GLBench

Change-Id: Ib8fc4b403638199f5e6390105d78dd5da4d1cab3
Reviewed-on: https://chromium-review.googlesource.com/202884
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>