Ping servo before waiting 60 seconds for ssh to timeout
There are some number of devices in the test lab that have DNS
names for servos but no actual servos.  When we run tests against
these hosts, we pay minutes of time just to wait for the SSH timeout
of 60 seconds when we try to ssh_ping() these servos to establish their
deadness.

BUG=chromium:380773
TEST=This makes wifitests against hosts as described above not take 5
minutes for a 30 second test.

Change-Id: I3759395dad15f54341b4312cb499db8991022af0
Reviewed-on: https://chromium-review.googlesource.com/204684
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>