Collect Settings Page screenshots.
We wish collect images for the settings page as we did for system tray.

Because the logic to collect these images is the same we pulled it up
to a parent class ui_test_base.

We applied the template pattern to define the screenshot collection
as the template method.

BUG=chromium:374994
TEST=Ran tests locally. Results can be seen at:
http://biopic.sandbox.google.com/5268094821859328/

Change-Id: Id421d15b06c37fb4b79babe2362a2f7a911b7890
Reviewed-on: https://chromium-review.googlesource.com/204800
Commit-Queue: Mussa Kiroga <mussa@google.com>
Tested-by: Mussa Kiroga <mussa@google.com>
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>