hardware_StorageFio: Add very quick workload for bvt suite
Add new workload that do random/sequential read/write over
64 MB area in spare root partition. The goal of this test
is to make sure that we don't break storage test.

The test is add to experimental suite for now and will be
moved over to bvt suite once it proved stable.

Runtime for this test is 5 seconds for machine with SSD
and less than a minute for machine with harddisk.

BUG=chromium:351900
TEST=Run with test_that and run perf_uploader_unittest

Change-Id: I448568a541242a65dda83a1290e996fba5196454
Reviewed-on: https://chromium-review.googlesource.com/205239
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>