platform_CrashStateful: Fix command execution error on panic reboot
The "run" command in reboot sometimes receives status 255 from ssh
indicating the connection was disconnected unexpectedly. For this test,
this is intentional because we are forcing a panic.

Wrap the provoke crash and breakme commands with redirects for stdout,
stderr, and stdin to and from /dev/null.

In addition, insert a sleep before the echo.

With these two changes, the ssh command should be given enough time
to close normally without error.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:387226
TEST=Run test_that <dut ip> platform_CrashStateful on skate,
squawks, and other platforms. Check the test passes without the
"reboot command failed" error.

Change-Id: Ic5b42aecd588616fadbd4e083566140b5052c016
Reviewed-on: https://chromium-review.googlesource.com/205681
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>