[autotest]: Add Moblab Host.
Because /usr/local/autotest is used by Moblab to be its main autotest
directory, we need client side testing for Moblab to not be installed
in the same directory.

This CL detects if the DUT is a Moblab Host, and if so returns
/tmp/autotest as the location to install autotest.

BUG=chromium:388067
TEST=Ran sleeptest via test_that and autotest was installed in /tmp/autotest,
verified that regular Cros devices still use /usr/local/autotest.

Change-Id: I4aaaf73568129ef57efb2124ed6de0300cbe64e4
Reviewed-on: https://chromium-review.googlesource.com/205827
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>