Remove firmware_programmer initialization.
Lab drones do not necessarily contain hdctools and the initialization of the
firmware programmer raises an exception. We need to move the programmer code
over to hctools, see chromium:281718.
Removing the initialization because nothing uses the functionality right now
and it stop causing an exception in the lab.

BUG=chromium:388945
TEST=Ran firmware_FAFTSetup and firmware_CgptStress on squawks at desk.

Change-Id: I6fa04ffe8d01bbc52c3df6ec8a19df00032b770d
Reviewed-on: https://chromium-review.googlesource.com/205955
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>