Re-enable firmware_programmer initialization.
The factory_flow tool in factory repo is temporarily relying on the
servo and firmware_programmer modules of autotest. The dependency
will no longer be needed when the logic is moved to hdctools.

This CL re-enables firmware_programmer with exception control, so
the initialization will not fail when hdctools is not installed.

BUG=chromium:388945
TEST=Ran firmware_FAFTSetup with the DUT on my desk.

Change-Id: I02eea566abbc64ed91567620a94e06d6e6663368
Reviewed-on: https://chromium-review.googlesource.com/206644
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>