power_suspend: Tweak whitelists for Parrot and Lumpy
This patch finally adds the 'ME failed to respond' error on Parrot to
the FirmwareError whitelist. We have seen these issues intermittently
for over a year, and there's no real interest in debugging it since the
ME is essentially a black box. We have tried to replace a few affected
DUTs, but it keeps cropping up and there's no real benefit in getting
reminded about this again. It doesn't seem to have any adverse effect on
the DUT.

Also expanded the existing FirmwareError whitelist mechanism on this
occasion to offer the same per-board granularity that the other
whitelists already have. This way we run less risk of overlooking new
errors (since firmware is often quite specific to a particular board
anyway).

On the good news side, it seems that the dreaded Lumpy touchpad bug is
actually resolved now (since after three months of cluelessness, we
finally got the helpful suggestion "have you tried turning it off and on
again?"). So it looks like for the first time in our history of sweeping
bugs under the rug in this test, we can actually remove a whitelist
entry again.

BUG=chromium:221538,chromium:288795
TEST=None

Change-Id: I6a06ebf9857e46813d3dff2a428c6f5eda63b316
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/207053
Reviewed-by: Sameer Nanda <snanda@chromium.org>