Install dependencies in additional steps firmware_SelfSignedBoot.
There were some places that autotest dependencies were not installed after
reboots. This is now resulting in BadStatusLine exceptions because the files
are missing when trying to start the RPCserver.

This seems to have started happening because of the way the new powerhandling
code implements the rec mode and reboot procedure.

BUG=None
TEST=Manual runs against wolf device

Change-Id: I9d338cd8f71099f71a418034f63ec9508fad91ab
Reviewed-on: https://chromium-review.googlesource.com/207066
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>