Update login_ChromeProfileSanitary for new Cookie DB behavior
It seems like Chrome no longer creates a cookie DB in the
default profile unless there are actually cookies there.
That's good, but this test was assuming the existence of
that file. Update the test to correctly verify that no DB
shows up. Maintain the check the verifies no change in a
DB as well, should one exist at the start of the test.

BUG=chromium:392510
TEST=run this test, it should pass.

Change-Id: Ib1b3f238c302170b50e71d523d085218d8d689c7
Reviewed-on: https://chromium-review.googlesource.com/207124
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>