Bring on tests for resolution switching videos.
We have image comparison based tests for video spanning mp4 and webm
formats, 720p and 1080p definitions.

We wish to add to the test suite a video that changes its resolution
from 720p to 1080p triggering the use of Media Source Extenstion (MSE).

With this change we have a different video so we are introducing a parameter
'video_name' that is specifiable from the control file.

BUG=chromium:393753
TEST=Ran local tests.

Change-Id: I5bafd710859f5d92dd9e439ce9b4505f6cfad42f
Reviewed-on: https://chromium-review.googlesource.com/207928
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>