Retry biopic upload if upload fails.
We have seen a number of test failures where the test got to the point
of uploading images but for some reason biopic server throws an exception.

We hypothesize that this is due the test uploading many images (~50)

This logic allows us to retry an upload upto a specified number incase of failure
to upload.

BUG=chromium:393682
TEST=Ran a sanity local to test to ensure we haven't broken anything. The real test
would be actual lab runs. We will be watching the dashboard to see a decrease in
BiopPic Upload related failures.

Change-Id: I68538dfa4cd79c2cc4e59f4d7f53400a8a9d7434
Reviewed-on: https://chromium-review.googlesource.com/207980
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>