Add the Python Imaging Library as an external dependency.
A recent change to support chameleon added a dependency on the
Python Imaging Library.  That library isn't guaranteed available on
all hosts that might run Autotest.  In particular, the library isn't
guaranteed to be present on the Autotest lab drone systems, which is
where it's most required.

This change adds the library using the Autotest external dependency
mechanism, and changes affected 'import' statements to account
for a location change.

BUG=None
TEST=manually test the import statements, plus pylint

Change-Id: I745a7d6050dfbf081447656213b2f9ed270d0a43
Reviewed-on: https://chromium-review.googlesource.com/209353
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>