Telemetry benchmarks run on battery power to report power data.
The Telemetry benchmarks recently changed to monitor power on CrOS when
executed on a DUT on battery power. This change attempts to turn power
off on a device before running the benchmarks in order to get the power
data added to the results. This also adds Telemetry's tab_switching
measurement to the lab to monitor idle browser tabs.

BUG=chromium:390699
TEST=Run tests.

Change-Id: I7914a3687502511a651410e4d033abd4600f1f76
Reviewed-on: https://chromium-review.googlesource.com/211204
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Tayler Mandel <tmandel@chromium.org>
Tested-by: Tayler Mandel <tmandel@chromium.org>