Add nyan_big SD and HD versions to test config.
nyan_big has a standard defition and a high definition version.
Accounting for that in tests.

Also there were few devices that were not added to the device_conf.
This caused the video_glitch_detection label to not be applied.
Fixing that.

BUG=chromium:393678
TEST=./autotest/files/contrib/print_host_labels.py <nyan_big_dut_ip>
Saw 'video_glitch_detection' in the result.

Change-Id: I290bbfbca2c9f5f7e2717214dea0af91bc66bcf3
Reviewed-on: https://chromium-review.googlesource.com/209561
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>