Test to compare installed and available firmware.
In the lab the DUTs all run test images.  That means that the firmware
updater is not automatically run.  This test does a comparison of the
BIOS and EC (if available) versions installed against the ones
available in the firmware shellball.

No software is installed.  This test is only to check the status of a
given DUT.

TEST=Manual; Ran agaist Peppy, Alex, and Zako
BUG=chromium:394517

Change-Id: I330884502a9591c7eb7f58492c2adcb8cd4c9801
Reviewed-on: https://chromium-review.googlesource.com/209679
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>